requirements.txt
mlflow==2.22.0
mlflow_skinny==2.22.0
optuna==4.3.0
pandas==2.2.3
PyYAML==6.0.2
PyYAML==6.0.2
scikit_learn==1.6.1
xgboost==3.0.2
ydata_profiling==4.16.1
